requirements.txt
aiohttp
charset-normalizer
click
dotenv
langchain
langchain-community
upstash-vector
tqdm
langchain-google-vertexai 
streamlit
langchain-openai